Many changes, read extended description
Distance calculation has been fixed and is now exact.
The Safe Zone is now a selectable POI, giving you directions away from the Storm.
All POIs are now clicked on when activating the script, placing a ping on their location. This gives extra stereo feedback on where the POI is, but also tells your teammates where you're going.
Removed unnecessary delays that were added previously.
ALT + S is no longer a keybind to activate Storm detection, as that is now baked into the POI detection.
Fixed an issue where the future-storm position outline was being detected as the player icon on the map.
If you are now within 75 meters of a POI, it says you are at the POI.
Selecting a POI from the POI list now clicks back into the Fortnite window automatically.