Fixed a few minor issues
Fixed an issue where the first button of the POI list would not be read properly by screen readers.
Fixed an issue where the placing of a marker on a POI would only work every other click due to the way pings work in Fortnite. This is fixed by removing the marker before placing another.